requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
jinja2==3.1.5
supabase==2.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==12.1.0
pandas==2.2.3
python-dotenv==1.0.1
httpx==0.28.1
